Bug fix/500/mobile view sytles fix (#560)
* Cancel modal mobile view fix

Fixed cancel sale modal styles.  Changed fixed size of width to have a maximum and a size relative if smaller than maximum.

* Fixed created by footer in sku detail

Fixed the created by footer in sku detail so whenever the screen is smaller the data wont be left out of the screen, and will be moved beneath.
Also fixed the information distance to the border for mobile devices.

* Modal outside of the screen

Solved "select payment to deposit" modal issues. In particular the modal size was bigger than its container, creating the modal outside the screen. Solved the issue by adding to the modal component a prop to determine the margin to the top of the page. Also fixed a minor issue regarding the payment method icons overlapping with the text