Bug fix/482/fix footer format layout (#569)
* Fix footer layout

Fixed the footer layout to match the design. Did a small refactor to match the form other components

* Updates for mobile view

Updated the styles to make mobile view match the design.

* Minimal change to avoid switch

Made a minimal change to avoid the use of switch.

* Fix chat icon overlap with footer

Override the styles of the hubspot widget to move it to a position which wont overlap with the footer. Probably a better solution can be found. Will discuss the magnitude of importance.
Fixed a little bug on mobile view of the  hedera logo and text.
Detected a bug corresponding to the chat icon staying on top of the drawer on mobile view.

* Modified line endings to match

Change line endings in new files from CRLF to LF to match prettier (visual studio code created the wrong line ending types by default)