Fixed styles and some bugs (#606)
Trying to replicate the issue find a bug which made the system prompt a message of "not logged in" which was an error in the implementation of the logic to check whether the user bid is high enough. While fixing this issue found an error in the way the conditions were evaluated that would leave to users being able to bid without enough funds.
The error related to the UI is happening since there is a conflict while using the styles overflow-y and backdrop-filter. Ended up taking out the backdrop-filter.
Made a change that was conflictive with develop branch, It is important it gets checked.
Particular changes:
-Change in History/index, on the function handleBid to make the code clearer and avoid having a prompt to login when there is a click on the button without being any bid.
-Created a style file for the modal, and took out the backdrop blur
-Imported the styles in Modal/index.tsx