Bug fix/528/bid takes too long (#613)
* Bid takes to long to be done

The bidding time linked to this issue depends directly on the time needed by the backend to resolve the API call, To make the user interface more friendly added a loader to show the user that the system is currently working.
Fixed along the way the logic behind bidding, related to minimum bid, price and a more user friendly error message which shows how much money is needed for his current bid.
Changed the timing of the toast when a bidding is successful to 5 seconds before reloading the page and also linked the FAQ instead of help.
Created a component loader which allows setting the color of the loader, height and other properties. Avoided changing the existing one to be sure no other components get broken.
Changes:
-Created Loader\index
-Modified history\index
-Modified Modal\Bid\index

* Added bid increment as variable

Bid increment used to be a fixed number =1,  now the number used is the on retrieved from backend.
Troubleshoot bug preventing a bid equal to the minimum, which was a backend issue.
Set one point of source to calculate the minimum bid to be used in other places.
Solved minor bug forcing the user to a higher bid.
particular changes:
-listings.ts: Added auction bid increment attribute.
-History\Index.tsx: Changed line to use the bidIncrement, but left a default value set to 1 in case some old data is not setting it (should be removed in future changes). Changed getPriceWithFee and getMinBid.

* Fixed error after conflict resolved

Fixed a prettier error after solving the conflict between branches.