Filter creators fix (#684)
There was a change in the naming of the properties of the user  which led to issues in the filters, by changing the details worked. Furthermore, fixed that the user was allow to select more than one creator to filter, which is not supported by the backend endpoint. If the user has selected a creator, then when selecting another one the last choice will be kept, and the other choice gets unselected.