Fixed mobile view footer getting out of place (#772)
Fixed the footer on the mobile view which was getting out of place on the landing page since there was a height in the main component that was fixed and was not working as expected. In order  to keep the code working in other cases the modification is only affecting smaller devices.