Delete Sku total (#897)
Deleted the Sku total from the view, which lead to numbers not aligned properly, which was fixed.
Also fixed the way the date string was wrapping to have the whole 'at 8:00 pm' moving down whenever necessary instead of moving just the 'pm'. Also diminished the padding right on the date to fix the way it was being shown on some resolutions.