Added a page loader to pagination change (#903)
Added a loader for pagination changes, before the user would touch on the pagination and see nothing until a few seconds loading, now the user is prompted with a loader. There are some calculations to define the height of the loader so the page won't resize when the loader is on.