Added search bar to collections (#957)
Added search bar to collections. Did not reuse the code of market place since many modifications were needed to adapt to this particular place. Created a mobile view which will be needed to be changed (there was no Figma for the case). Added message corresponding to empty search results. Created missing styles files.